set pattern, debounce and rename entity editor
